+++ src/commands/CreateCustomizerCommand.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import { ConnectorsTreeItem } from "../models/TreeModel";
+import { CustomizersTreeItem } from "../models/TreeModel";
 import { SaaSConnectivityClientFactory } from "../services/SaaSConnectivityClientFactory";
 import { ISCExtensionClient } from "../iscextension/iscextension-client";
 import { isEmpty } from "../utils/stringUtils";
@@ -9,9 +9,9 @@ import { confirm } from "../utils/vsCodeHelpers";
 export async function askName(tenantDisplayName: string): Promise<string | undefined> {
     const result = await vscode.window.showInputBox({
         ignoreFocusOut: true,
-        placeHolder: 'alias',
-        prompt: "Enter a name for the connector",
-        title: `Create Connector In ${tenantDisplayName}`,
+        placeHolder: 'name',
+        prompt: "Enter a name for the customizer",
+        title: `Create Customizer In ${tenantDisplayName}`,
         validateInput: text => {
             if (isEmpty(text) || isEmpty(text.trim())) {
                 return "Name must not be empty";
@@ -21,7 +21,7 @@ export async function askName(tenantDisplayName: string): Promise<string | undef
     return result;
 }
 
-export class CreateConnectorCommand {
+export class CreateCustomizerCommand {
     private readonly factory: SaaSConnectivityClientFactory
     private readonly iscExtensionClient: ISCExtensionClient
 
@@ -30,25 +30,24 @@ export class CreateConnectorCommand {
         this.factory = new SaaSConnectivityClientFactory(this.iscExtensionClient)
     }
 
-    public async execute(item: ConnectorsTreeItem) {
+    public async execute(item: CustomizersTreeItem) {
 
         const isReadOnly = this.iscExtensionClient.isTenantReadonly(item.tenantId)
 
-        if (isReadOnly && !(await confirm(`The tenant ${item.tenantDisplayName} is read-only. Do you still want to create a connector?`))) {
+        if (isReadOnly && !(await confirm(`The tenant ${item.tenantDisplayName} is read-only. Do you still want to create a customizer?`))) {
             return
         }
 
-        const alias = await askName(item.tenantDisplayName)
-        if (alias === undefined) { return }
+        const name = await askName(item.tenantDisplayName)
+        if (name === undefined) { return }
 
         const client = await this.factory.getSaaSConnectivityClient(item.tenantId, item.tenantName)
         try {
-            const connector = await client.createConnector(alias)
-            vscode.window.showInformationMessage(`Connector ${connector.alias} created with id ${connector.id}`)
+            const customizer = await client.createCustomizer(name)
+            vscode.window.showInformationMessage(`Customizer ${customizer.name} created with id ${customizer.id}`)
             vscode.commands.executeCommand(constants.REFRESH, item);
-
         } catch (error) {
-            vscode.window.showErrorMessage(`Could not create connector: ${error}`)
+            vscode.window.showErrorMessage(`Could not create customizer: ${error}`)
         }
     }