+++ src/models/API.ts
@@ -68,3 +68,50 @@ export interface GetInstancesResponse {
     connectorCustomizerId: string;
 }
 
+
+export interface CreateCustomizerRequest {
+    name: string
+}
+
+export interface CustomizerVersion {
+    connectorId: string
+    version: number
+}
+
+export interface Customizer {
+    id: string;
+    name: string;
+    imageVersion: string;
+}
+
+export interface Instance {
+    id: string;
+    name: string;
+    connectorCustomizerId: string;
+}
+
+export interface CreateCustomizerResponse extends Customizer {
+}
+
+export interface UpdateCustomizerRequest {
+    name: string
+}
+
+export interface UpdateCustomizerResponse extends Customizer { }
+
+export interface UploadCustomizerResponse extends CustomizerVersion { }
+
+
+export interface GetInstancesResponse extends Instance { }
+
+export interface LinkRequest {
+    op: string;
+    path: string;
+    value?: string;
+}
+
+export interface LinkResponse extends Instance { }
+
+export interface UnlinkRequest extends LinkRequest { }
+
+export interface UnlinkResponse extends LinkResponse { }